+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [v1.3.1](https://github.com/cacing69/cquery/compare/v1.3.1...v1.3.0) (2023-09-0?)
 
 - Add method pluck on collection for get single array from collection given
+- Add method last on cquery for get single array from collection given
 
 ## [v1.3.0](https://github.com/cacing69/cquery/compare/v1.2.0...v1.3.0) (2023-09-09)